Grading
Great app!

Interesting choices for plots, but it doesn't matter that much because they work and are reactive!

Only real issue is the rownames not being taken out when loading or set to false in the write.csv() function in the app, so when I download the data I end up getting two rowname columns